requirements.txt
asgiref==3.6.0
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
cffi==1.15.1
channels==4.0.0
constantly==15.1.0
cryptography==39.0.2
daphne==4.0.0
Django==4.1.7
djangorestframework==3.14.0
generics==6.0.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pytz==2022.7.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
zope.interface==6.0
